refactor(api): split monolithic api.ts into per-feature modules
Break the ~725-line api.ts into 21 focused files under services/api/:
- apiClient.ts for shared utilities (fetchWithAuth, API_BASE_URL)
- One *.api.ts file per domain (players, matches, championships, etc.)
- Barrel index.ts re-exports everything, preserving all import paths
