+++ frontend/src/services/api/apiClient.ts
@@ -1,0 +1,37 @@
+export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://localhost:3001';
+
+// Helper to get auth token from session storage (uses Cognito access token)
+export const getAuthToken = (): string | null => {
+  return sessionStorage.getItem('accessToken');
+};
+
+// Helper to make authenticated requests with optional abort signal
+export const fetchWithAuth = async (url: string, options: RequestInit = {}, signal?: AbortSignal) => {
+  const token = getAuthToken();
+  const headers: Record<string, string> = {
+    'Content-Type': 'application/json',
+    ...(options.headers as Record<string, string> || {}),
+  };
+
+  if (token) {
+    headers['Authorization'] = `Bearer ${token}`;
+  }
+
+  const response = await fetch(url, {
+    ...options,
+    headers,
+    signal,
+  });
+
+  if (!response.ok) {
+    const error = await response.json().catch(() => ({ message: 'Request failed' }));
+    throw new Error(error.message || `HTTP ${response.status}`);
+  }
+
+  // Handle 204 No Content responses (e.g., from DELETE operations)
+  if (response.status === 204) {
+    return undefined;
+  }
+
+  return response.json();
+};
