+++ frontend/src/services/api/auth.api.ts
@@ -1,0 +1,20 @@
+import { getAuthToken } from './apiClient';
+
+export const authApi = {
+  setToken: (token: string) => {
+    sessionStorage.setItem('accessToken', token);
+  },
+
+  clearToken: () => {
+    sessionStorage.removeItem('accessToken');
+    sessionStorage.removeItem('idToken');
+  },
+
+  isAuthenticated: (): boolean => {
+    return !!getAuthToken();
+  },
+
+  getToken: (): string | null => {
+    return getAuthToken();
+  },
+};
