+++ frontend/src/services/api/players.api.ts
@@ -1,0 +1,28 @@
+import type { Player } from '../../types';
+import { API_BASE_URL, fetchWithAuth } from './apiClient';
+
+export const playersApi = {
+  getAll: async (signal?: AbortSignal): Promise<Player[]> => {
+    return fetchWithAuth(`${API_BASE_URL}/players`, {}, signal);
+  },
+
+  create: async (player: Omit<Player, 'playerId' | 'createdAt' | 'updatedAt'>): Promise<Player> => {
+    return fetchWithAuth(`${API_BASE_URL}/players`, {
+      method: 'POST',
+      body: JSON.stringify(player),
+    });
+  },
+
+  update: async (playerId: string, updates: Partial<Player>): Promise<Player> => {
+    return fetchWithAuth(`${API_BASE_URL}/players/${playerId}`, {
+      method: 'PUT',
+      body: JSON.stringify(updates),
+    });
+  },
+
+  delete: async (playerId: string): Promise<void> => {
+    return fetchWithAuth(`${API_BASE_URL}/players/${playerId}`, {
+      method: 'DELETE',
+    });
+  },
+};
